pyproject.toml
[tool.poetry]
name = "cloud-watch-test"
version = "0.1.0"
description = ""
authors = ["chengying <chengying@choirai.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
flask = "^3.0.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
